fix: 🐛 fix test cases
