+++ app/app_test.go
@@ -222,7 +222,7 @@ func TestSimAppExportAndBlockedAddrs(t *testing.T) {
 	encCfg := MakeTestEncodingConfig()
 	db := dbm.NewMemDB()
 	app := NewSifApp(log.NewTMLogger(log.NewSyncWriter(os.Stdout)), db, nil, true, map[int64]bool{}, DefaultNodeHome, 0, encCfg, EmptyAppOptions{})
-
+	SetConfig(false)
 	for acc := range maccPerms {
 		require.True(
 			t,
