+++ proto/sifnode/clp/v1/types.proto
@@ -8,62 +8,6 @@ option go_package = "github.com/Sifchain/sifnode/x/clp/types";
 
 message Asset { string symbol = 1; }
 
-message Pool {
-  Asset external_asset = 1;
-  string native_asset_balance = 2 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Uint",
-    (gogoproto.nullable) = false,
-    (gogoproto.moretags) = "yaml:\"native_asset_balance\""
-  ];
-  string external_asset_balance = 3 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Uint",
-    (gogoproto.nullable) = false,
-    (gogoproto.moretags) = "yaml:\"external_asset_balance\""
-  ];
-  string pool_units = 4 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Uint",
-    (gogoproto.nullable) = false,
-    (gogoproto.moretags) = "yaml:\"pool_units\""
-  ];
-    string swap_price_native = 5 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Dec",
-    (gogoproto.moretags) = "yaml:\"swap_price_native \""
-  ];
-  string swap_price_external = 6 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Dec",
-    (gogoproto.moretags) = "yaml:\"swap_price_external \""
-  ];
-  string reward_period_native_distributed = 7 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Uint",
-    (gogoproto.nullable) = false,
-    (gogoproto.moretags) = "yaml:\"reward_period_native_distributed\""
-  ];
-  string external_liabilities = 8 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Uint",
-    (gogoproto.nullable) = false
-  ];
-  string external_custody = 9 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Uint",
-    (gogoproto.nullable) = false
-  ];
-  string native_liabilities = 10 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Uint",
-    (gogoproto.nullable) = false
-  ];
-  string native_custody = 11 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Uint",
-    (gogoproto.nullable) = false
-  ];
-  string health = 12 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Dec",
-    (gogoproto.nullable) = false
-  ];
-  string interest_rate = 13 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Dec",
-    (gogoproto.nullable) = false
-  ];
-}
-
 message LiquidityProvider {
   Asset asset = 1;
   string liquidity_provider_units = 2 [
