+++ x/margin/keeper/genesis_test.go
@@ -38,15 +38,21 @@ func TestKeeper_ExportGenesis(t *testing.T) {
 		require.NotNil(t, marginKeeper)
 
 		params := types.Params{
-			LeverageMax:           sdk.NewDec(10),
-			InterestRateMax:       sdk.NewDec(5),
-			InterestRateMin:       sdk.NewDec(1),
-			InterestRateIncrease:  sdk.NewDec(1),
-			InterestRateDecrease:  sdk.NewDec(1),
-			HealthGainFactor:      sdk.NewDec(1),
-			EpochLength:           1,
-			ForceCloseThreshold:   sdk.NewDec(1), //TODO get real default
-			RemovalQueueThreshold: sdk.ZeroDec(),
+			LeverageMax:              sdk.NewDec(2),
+			HealthGainFactor:         sdk.NewDec(1),
+			InterestRateMin:          sdk.NewDecWithPrec(5, 3),
+			InterestRateMax:          sdk.NewDec(3),
+			InterestRateDecrease:     sdk.NewDecWithPrec(1, 1),
+			InterestRateIncrease:     sdk.NewDecWithPrec(1, 1),
+			ForceCloseThreshold:      sdk.NewDecWithPrec(1, 1),
+			ForceCloseFundPercentage: sdk.NewDecWithPrec(1, 1),
+			InsuranceFundAddress:     "sif1syavy2npfyt9tcncdtsdzf7kny9lh777yqc2nd",
+			PoolOpenThreshold:        sdk.NewDecWithPrec(1, 1),
+			RemovalQueueThreshold:    sdk.NewDecWithPrec(1, 1),
+			EpochLength:              1,
+			MaxOpenPositions:         10000,
+			SqModifier:               sdk.MustNewDecFromStr("10000000000000000000000000"),
+			SafetyFactor:             sdk.MustNewDecFromStr("1.05"),
 		}
 		want := types.GenesisState{Params: &params}
 		marginKeeper.InitGenesis(ctx, want)
@@ -58,12 +64,29 @@ func TestKeeper_ExportGenesis(t *testing.T) {
 }
 
 func TestKeeper_InitGenesis(t *testing.T) {
-	t.Run("params with empty fields", func(t *testing.T) {
+	t.Run("params with initial fields", func(t *testing.T) {
 		ctx, app := test.CreateTestAppMargin(false)
 		marginKeeper := app.MarginKeeper
 		require.NotNil(t, marginKeeper)
 
-		params := types.Params{}
+		params := types.Params{
+			LeverageMax:              sdk.NewDec(2),
+			HealthGainFactor:         sdk.NewDec(1),
+			InterestRateMin:          sdk.NewDecWithPrec(5, 3),
+			InterestRateMax:          sdk.NewDec(3),
+			InterestRateDecrease:     sdk.NewDecWithPrec(1, 1),
+			InterestRateIncrease:     sdk.NewDecWithPrec(1, 1),
+			ForceCloseThreshold:      sdk.NewDecWithPrec(1, 1),
+			ForceCloseFundPercentage: sdk.NewDecWithPrec(1, 1),
+			InsuranceFundAddress:     "sif1syavy2npfyt9tcncdtsdzf7kny9lh777yqc2nd",
+			PoolOpenThreshold:        sdk.NewDecWithPrec(1, 1),
+			RemovalQueueThreshold:    sdk.NewDecWithPrec(1, 1),
+			EpochLength:              1,
+			MaxOpenPositions:         10000,
+			Pools:                    []string{},
+			SqModifier:               sdk.MustNewDecFromStr("10000000000000000000000000"),
+			SafetyFactor:             sdk.MustNewDecFromStr("1.05"),
+		}
 		want := types.GenesisState{Params: &params}
 		validatorUpdate := marginKeeper.InitGenesis(ctx, want)
 
@@ -87,15 +110,21 @@ func TestKeeper_InitGenesis(t *testing.T) {
 		require.NotNil(t, marginKeeper)
 
 		params := types.Params{
-			LeverageMax:           sdk.NewDec(10),
-			InterestRateMax:       sdk.NewDec(5),
-			InterestRateMin:       sdk.NewDec(1),
-			InterestRateIncrease:  sdk.NewDec(1),
-			InterestRateDecrease:  sdk.NewDec(1),
-			HealthGainFactor:      sdk.NewDec(1),
-			EpochLength:           1,
-			ForceCloseThreshold:   sdk.NewDec(1), //TODO get real default
-			RemovalQueueThreshold: sdk.ZeroDec(),
+			LeverageMax:              sdk.NewDec(2),
+			HealthGainFactor:         sdk.NewDec(1),
+			InterestRateMin:          sdk.NewDecWithPrec(5, 3),
+			InterestRateMax:          sdk.NewDec(3),
+			InterestRateDecrease:     sdk.NewDecWithPrec(1, 1),
+			InterestRateIncrease:     sdk.NewDecWithPrec(1, 1),
+			ForceCloseThreshold:      sdk.NewDecWithPrec(1, 1),
+			ForceCloseFundPercentage: sdk.NewDecWithPrec(1, 1),
+			InsuranceFundAddress:     "sif1syavy2npfyt9tcncdtsdzf7kny9lh777yqc2nd",
+			PoolOpenThreshold:        sdk.NewDecWithPrec(1, 1),
+			RemovalQueueThreshold:    sdk.NewDecWithPrec(1, 1),
+			EpochLength:              1,
+			MaxOpenPositions:         10000,
+			SqModifier:               sdk.MustNewDecFromStr("10000000000000000000000000"),
+			SafetyFactor:             sdk.MustNewDecFromStr("1.05"),
 		}
 		want := types.GenesisState{Params: &params}
 		validatorUpdate := marginKeeper.InitGenesis(ctx, want)
